requirements.txt
httpx==0.28.1
beautifulsoup4==4.13.4
psycopg2==2.9.10
python-dotenv==1.1.1